+++ src/app.module.ts
@@ -16,6 +16,7 @@ import { ChangelogModule } from '@src/common/changelog/changelog.module';
 import { RoomsModule } from './modules/rooms/rooms.module';
 import { MediaModule } from './modules/media/media.module';
 import { RandomMatchModule } from './modules/random-match/random-match.module';
+import { ContactsModule } from './modules/contacts/contacts.module';
 import { APP_GUARD } from '@nestjs/core';
 import { JwtAuthGuard } from '@src/common/guards/jwt-auth.guard';
 import { PermissionGuard } from '@src/common/guards/permission.guard';
@@ -41,6 +42,7 @@ import { TenantGuard } from '@src/common/guards/tenant.guard';
     InterestsModule,
     MailModule,
     RoomsModule,
+    ContactsModule,
     MediaModule,
     RandomMatchModule,
     HealthModule,