+++ src/modules/contacts/dto/create-contact.response.dto.ts
@@ -1,0 +1,8 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class CreateContactResponseDto {
+  @ApiProperty({
+    example: true,
+  })
+  success: true;
+}