+++ src/modules/contacts/swagger-docs/index.ts
@@ -1,0 +1,2 @@
+export * from './save-contact';
+export * from './remove-contact';