+++ src/modules/users/dto/updated-user.dto.ts
@@ -72,6 +72,12 @@ export class UpdatedUserDto {
   })
   gender?: Gender;
 
+  @ApiPropertyOptional({
+    example: true,
+    description: 'Allow other users to save this profile in contacts',
+  })
+  allowContactSave?: boolean;
+
   @ApiProperty({
     enum: AuthProvider,
     example: AuthProvider.LOCAL,